hashing stuff
Former-commit-id: f89389860122c859b0decfcd1888583888eca496

+++ src/gmap.cpp
@@ -2,11 +2,11 @@
 // Needs a relatively high count to work with zero collisions..
 // Is this to be expected? Or can we find a better hash function
 // or possibly find another solution?
-inline u32 HashString(u32 Size, const void* Key)
+inline u64 HashString(u64 Size, const void* Key)
 {
-    unsigned long Hash = 5381;
+    u64 Hash = 5381;
     char* K = (char*)Key;
-    u32 C;
+    u64 C;
     
     while (C = *K++)
         Hash = ((Hash << 5) + Hash) + C; /* hash * 33 + c */
@@ -16,12 +16,12 @@ inline u32 HashString(u32 Size, const void* Key)
 
 // Jenkins one at a time hash
 // NOTE(niels): https://en.wikipedia.org/wiki/Jenkins_hash_function
-inline u32 HashStringJenkins(u32 Size, const void* Key) 
+inline u64 HashStringJenkins(u64 Size, const void* Key) 
 {
     char* K = (char*)Key;
-    u32 Length = (u32)strlen(K);
-    u32 I = 0;
-    u32 Hash = 0;
+    u64 Length = (u64)strlen(K);
+    u64 I = 0;
+    u64 Hash = 0;
     while (I != Length) {
         Hash += K[I++];
         Hash += Hash << 10;
@@ -33,8 +33,8 @@ inline u32 HashStringJenkins(u32 Size, const void* Key)
     return Hash & (Size - 1);
 }
 
-inline u32 HashInt(u32 Size, const void* Key) {
-    u32 K = *(u32*)(Key);
+inline u64 HashInt(u64 Size, const void* Key) {
+    u64 K = *(u64*)(Key);
     K = ((K >> 16) ^ K) * 0x45d9f3b;
     K = ((K >> 16) ^ K) * 0x45d9f3b;
     K = (K >> 16) ^ K;
@@ -54,7 +54,7 @@ u32 HashInt(u32 Size, const void* A)
     return K & (Size - 1);
 }*/
 
-inline u32 HashIntKeys(u32 Size, const void* Key)
+inline u64 HashIntKeys(u64 Size, const void* Key)
 {
     u32 K = *(u32*)(Key);
     return K % (Size - 1);
